Improve parser for better rendering of index.html file (#60)
* skip unidentified tags in handle_starttag

Here we make the compiler to skip any unidentified tag that is not found in the generic tags or custom components

* Add error handling for unidentified tags in parser

This commit adds error handling to the parser to detect and handle unidentified tags during the parsing process. If an unknown tag is encountered, the parser now throws an UnknownTagError. The error message includes the name of the unrecognized tag and provides guidance to the developer to review the specific line in their pyml code where the issue occurred. This enhancement improves the developer experience by promptly identifying and reporting any unidentified tags, helping them identify and resolve the issue more efficiently

* improve error-handler for parser

changed generic_tags from list to a set datatype. Rename 'class_name' to avoid potential conflict. Ensure that test-application is error-free