feat(component.py,-dom_method,py,-parser.py,-parent.fyre,-store.fyre): Added special tag <slot> which specify position of inner content of it's parent component
In component class I defined new list to store inner content of special tag <slot>. I also defined a variable called is_costum "meaning that is our <slot> tag"  to be True. In parser.py I check if the parent node is our special tag '<slot>', and if it's true, the current component is appended to the inner content list instead of the children list. Then in the dom_methods.py I render the component properties of inner_content list. For every child in children list, I check if the child tag is our <slot>, and if it's true, I added the <slot> component inner content to its own properties.

BREAKING CHANGE: this closes issue 39