fix(dom_methods.py): handle a edge case when specifying position of child
Implemented code with a variable is_slot_used = False. Then if childrenElement.tag is the <slot>, I set is_slot_used equals True. Then if the <slot> tag is not used in the component definition but we have content in inner_html, that means the user forgot "or is not aware" that he should specify where the children content should be added. Based on that I raise an error to let user knows that he is missing the <slot> tag.